Move to separate file. Ready to go.
